requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.30.0
hyperliquid-python-sdk>=0.9.0
eth-account>=0.13.0
pydantic>=2.0
python-dotenv>=1.0.0
rich>=13.0
